+++ src/accessory/JsonParse.h
@@ -1,8 +1,8 @@
 #pragma once
 
-long long StrToDecimal(const string_view& svData, int power);
-long long StrToDecimal2(const string_view& svData, int power);
-long Str2Long(const std::string_view& svData, int power);
+long long StrToDecimal(const string_view& svData, size_t power);
+long long StrToDecimal2(const string_view& svData, size_t power);
+long Str2Long(const std::string_view& svData, size_t power);
 string_view GetNextField(const string_view& svData, size_t& lCurrentPos, char delimiter);
 
 void ReportJSonErrorToSystemMessage(const string& strPrefix, const string& strWhat);
